bugFix(fixed bugs)
-fixed typescript errors
-cleaned up code
