+++ backend/controllers/auth.controller.js
@@ -91,7 +91,6 @@ export const verifyEmail = async (req, res) => {
 			},
 		});
 	} catch (error) {
-		console.log("error in verifyEmail ", error);
 		res.status(500).json({ success: false, message: "Server error" });
 	}
 };
@@ -122,7 +121,6 @@ export const login = async (req, res) => {
 			},
 		});
 	} catch (error) {
-		console.log("Error in login ", error);
 		res.status(400).json({ success: false, message: error.message });
 	}
 };
@@ -155,7 +153,6 @@ export const forgotPassword = async (req, res) => {
 
 		res.status(200).json({ success: true, message: "Password reset link sent to your email" });
 	} catch (error) {
-		console.log("Error in forgotPassword ", error);
 		res.status(400).json({ success: false, message: error.message });
 	}
 };
@@ -186,7 +183,6 @@ export const resetPassword = async (req, res) => {
 
 		res.status(200).json({ success: true, message: "Password reset successful" });
 	} catch (error) {
-		console.log("Error in resetPassword ", error);
 		res.status(400).json({ success: false, message: error.message });
 	}
 };
@@ -200,7 +196,6 @@ export const checkAuth = async (req, res) => {
 
 		res.status(200).json({ success: true, user });
 	} catch (error) {
-		console.log("Error in checkAuth ", error);
 		res.status(400).json({ success: false, message: error.message });
 	}
 };
@@ -211,7 +206,6 @@ export const getUsers = async (req, res) => {
 		const users = await User.find().select("-password");
 		res.status(200).json({ success: true, users });
 	} catch (error) {
-		console.log("Error in getUsers ", error);
 		res.status(400).json({ success: false, message: error.message });
 	}
 };
