+++ backend/mail/emails.js
@@ -17,7 +17,6 @@ export const sendVerificationEmail = async (email, verificationToken) => {
 			category: "Email Verification",
 		});
 
-		console.log("Email sent successfully", response);
 	} catch (error) {
 		console.error(`Error sending verification`, error);
 
@@ -39,7 +38,6 @@ export const sendWelcomeEmail = async (email, name) => {
 			},
 		});
 
-		console.log("Welcome email sent successfully", response);
 	} catch (error) {
 		console.error(`Error sending welcome email`, error);
 
@@ -59,8 +57,6 @@ export const sendPasswordResetEmail = async (email, resetURL) => {
 			category: "Password Reset",
 		});
 	} catch (error) {
-		console.error(`Error sending password reset email`, error);
-
 		throw new Error(`Error sending password reset email: ${error}`);
 	}
 };
@@ -77,10 +73,7 @@ export const sendResetSuccessEmail = async (email) => {
 			category: "Password Reset",
 		});
 
-		console.log("Password reset email sent successfully", response);
 	} catch (error) {
-		console.error(`Error sending password reset success email`, error);
-
 		throw new Error(`Error sending password reset success email: ${error}`);
 	}
 };
