+++ backend/middleware/verifyToken.js
@@ -14,7 +14,6 @@ export const verifyToken = (req, res, next) => {
 		req.userId = decoded.userId;
 		next();
 	} catch (error) {
-		console.log("Error in verifyToken ", error);
 		return res.status(500).json({ success: false, message: "Server error" });
 	}
 };
