+++ backend/package.json
@@ -5,7 +5,8 @@
   "type": "module",
   "scripts": {
     "start": "nodemon index.js",
-    "build": "echo 'No build needed'"
+    "build": "echo 'No build needed'",
+    "deploy:fly": "flyctl deploy --app projectmanagementool"
   },
   "keywords": [],
   "author": "",
