+++ frontend/src/app/login/page.tsx
@@ -57,7 +57,6 @@ export default function LoginPage() {
       }
       // eslint-disable-next-line @typescript-eslint/no-explicit-any
     } catch (error: any) {
-      console.error("Login error:", error);
       toast.error(
         error.response?.data?.message ||
           error.message ||
